--- lib/sccm/win32-certstore/win32/certstore/version.rb
@@ -1,6 +1,0 @@
-module Win32
-  class Certstore
-    VERSION = "0.6.3".freeze
-    MAJOR, MINOR, TINY = VERSION.split(".")
-  end
-end
